Created ENASubmittable interface and moved existing ENASubmittable abstract class to AbstractENASubmittable
