+++ subs-ena-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/BaseSubmittableFactory.java
@@ -3,14 +3,25 @@ package uk.ac.ebi.subs.data.submittable;
 /**
  * Created by neilg on 05/04/2017.
  */
-public interface BaseSubmittableFactory<T extends Submittable>   {
-    T getBaseObject ();
-    void setBaseSubmittable(BaseSubmittable baseSubmittable) throws IllegalAccessException;
+public interface BaseSubmittableFactory<T extends Submittable> extends ENASubmittable<T> {
+    void setBaseSubmittable(Submittable submittable) throws IllegalAccessException;
+    Submittable createNewSubmittable();
 
-    static BaseSubmittableFactory create (Class<? extends BaseSubmittableFactory> clasz,
-                                          BaseSubmittable baseSubmittable) throws IllegalAccessException, InstantiationException {
+    static ENASubmittable create (Class<? extends BaseSubmittableFactory> clasz,
+                                  Submittable submittable) throws IllegalAccessException, InstantiationException {
         BaseSubmittableFactory baseSubmittableFactory = clasz.newInstance();
-        baseSubmittableFactory.setBaseSubmittable(baseSubmittable);
+        baseSubmittableFactory.setBaseSubmittable(submittable);
         return baseSubmittableFactory;
     }
+
+    static ENASubmittable create (Class<? extends BaseSubmittableFactory> clasz) throws IllegalAccessException, InstantiationException {
+        BaseSubmittableFactory baseSubmittableFactory = clasz.newInstance();
+        final Submittable newSubmittable = baseSubmittableFactory.createNewSubmittable();
+        return create(clasz,newSubmittable);
+    }
+
+
+
+
+
 }
