+++ subs-ena-data-model/src/main/java/uk/ac/ebi/subs/data/submittable/ENAStudy.java
@@ -1,29 +1,21 @@
 package uk.ac.ebi.subs.data.submittable;
 
-import org.springframework.beans.BeanUtils;
-
-import uk.ac.ebi.subs.data.component.Attribute;
-import uk.ac.ebi.subs.data.component.Contacts;
-import uk.ac.ebi.subs.data.component.Publications;
-import uk.ac.ebi.subs.data.component.Team;
 import uk.ac.ebi.subs.ena.annotation.ENAAttribute;
 import uk.ac.ebi.subs.ena.annotation.ENAValidation;
 
-import java.util.List;
-
-@ENAValidation(requiredAttributes = {"existing_study_type"})
-public class ENAStudy extends ENASubmittable<Study> {
+@ENAValidation(requiredAttributes = {"existing_study_type","study_abstract"})
+public class ENAStudy extends AbstractENASubmittable<Study> {
 
     public static final String EXISTING_STUDY_TYPE = "existing_study_type";
     @ENAAttribute(name = EXISTING_STUDY_TYPE)
     String studyType;
-    String description;
+    public static final String STUDY_ABSTRACT = "study_abstract";
+    @ENAAttribute(name = STUDY_ABSTRACT)
+    String studyAbstract;
 
     public ENAStudy (Study study) throws IllegalAccessException {
         super(study);
-        //BeanUtils.copyProperties(study,this);
         //serialiseAttributes();
-        final BaseSubmittable<Study> baseSubmittable = getBaseSubmittable();
     }
 
     public ENAStudy () throws IllegalAccessException {
@@ -39,4 +31,16 @@ public class ENAStudy extends ENASubmittable<Study> {
         this.studyType = studyType;
     }
 
+    public String getStudyAbstract() {
+        return studyAbstract;
+    }
+
+    public void setStudyAbstract(String studyAbstract) {
+        this.studyAbstract = studyAbstract;
+    }
+
+    @Override
+    public Submittable createNewSubmittable() {
+        return new Study();
+    }
 }
