+++ subs-ena-data-model/src/main/resources/uk/ac/ebi/subs/data/submittable/study_mapping.xml
@@ -6,7 +6,7 @@
     <java-types>
         <java-type name="Submittable" xml-transient="true"/>
         <java-type name="BaseSubmittable" xml-transient="true"/>
-        <java-type name="ENASubmittable" xml-transient="true"/>
+        <java-type name="AbstractENASubmittable" xml-transient="true"/>
         <!--<java-type name="ENASubmittable" xml-transient="true">
             <java-attributes>
                 <xml-attribute java-attribute="accession"/>
@@ -20,7 +20,7 @@
         <java-type name="Study" xml-transient="true"/>
         <java-type name="ENAStudy" xml-accessor-type="NONE">
             <xml-root-element name="STUDY"/>
-            <xml-type prop-order="accession alias teamName title studyType description attributes"/>
+            <xml-type prop-order="accession alias teamName title studyType studyAbstract description attributes"/>
             <java-attributes>
                 <xml-attribute java-attribute="accession"/>
                 <xml-attribute java-attribute="alias"/>
@@ -30,6 +30,7 @@
                 <xml-attribute java-attribute="teamName" xml-path="IDENTIFIERS/SUBMITTER_ID/@namespace"/>
                 <xml-element java-attribute="title" xml-path="DESCRIPTOR/STUDY_TITLE/text()"/>
                 <xml-element java-attribute="studyType" xml-path="DESCRIPTOR/STUDY_TYPE/@existing_study_type"/>
+                <xml-element java-attribute="studyAbstract" xml-path="DESCRIPTOR/STUDY_ABSTRACT/text()"/>
                 <xml-element java-attribute="description" xml-path="DESCRIPTOR/STUDY_DESCRIPTION/text()"/>
                 <xml-element java-attribute="attributes" name="STUDY_ATTRIBUTE"
                              type="uk.ac.ebi.subs.data.component.Attribute" container-type="java.util.ArrayList">
